Move MetricConfiguration to separate file.
